app/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
python-multipart
tenacity

# Gemini (new SDK - replaces deprecated google-generativeai)
google-genai

# CurricuLLM (OpenAI-compatible API)
openai

# Firebase
firebase-admin

# Document parsing (rubric file uploads)
PyPDF2
python-docx

# Keep PostgreSQL support for local Docker dev (Firestore used in production)
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
google-generativeai
python-multipart
tenacity
python-dotenv
